+++ custom_components/tuya_ble/switch.py
@@ -1132,4 +1132,7 @@ async def async_setup_entry(
                     mapping,
                 )
             )
+    from .schema_entities import schema_entities
+
+    entities.extend(schema_entities(hass, data, Platform.SWITCH))
     async_add_entities(entities)